# docs/swagger.yaml
basePath: /api
definitions:
  auth.loginResponse:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱账号
        type: string
      expire_time:
        description: token过期时间
        type: integer
      id:
        description: 用户id
        type: integer
      name:
        description: 名称
        type: string
      token:
        description: token
        type: string
      ttl:
        description: token有效期
        type: integer
      uid:
        description: uid
        type: string
    type: object
  friend.UserStatus:
    properties:
      id:
        description: 用户id
        type: integer
      status:
        description: 0 未在线 1 在线
        type: integer
    type: object
  im_friend_records.ImFriendRecords:
    properties:
      created_at:
        type: string
      form_id:
        type: integer
      id:
        type: integer
      information:
        description: 请求信息
        type: string
      status:
        description: 0 等待通过 1 已通过 2 已拒绝
        type: integer
      to_id:
        type: integer
      users:
        $ref: '#/definitions/im_friend_records.ImUsers'
    type: object
  im_friend_records.ImUsers:
    properties:
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  im_friends.ImFriends:
    properties:
      created_at:
        type: string
      form_id:
        type: integer
      id:
        type: integer
      note:
        type: string
      status:
        description: 0.未置顶 1.已置顶
        type: integer
      to_id:
        type: integer
      top_time:
        type: string
      uid:
        type: string
      updated_at:
        type: string
      users:
        $ref: '#/definitions/im_friends.ImUsers'
    type: object
  im_friends.ImUsers:
    properties:
      age:
        type: integer
      avatar:
        type: string
      bio:
        type: string
      client_type:
        type: integer
      email:
        type: string
      id:
        type: integer
      last_login_time:
        type: string
      name:
        type: string
      sex:
        type: integer
      status:
        type: integer
      uid:
        type: string
    type: object
...
  response.JsonResponse:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 返回结构数据
      httpCode:
        description: http状态码
        type: integer
      message:
        description: 消息内容
        type: string
    type: object
...
host: 127.0.0.1:8000
info:
  contact:
    email: pltrueover@gmail.com
    name: im-services
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: im-services  接口文档
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - multipart/form-data
      description: 登录接口
      parameters:
      - description: 账号
        in: formData
        name: email
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.loginResponse'
              type: object
      summary: Login 登录
      tags:
      - 登录相关
  /auth/registered:
    post:
      consumes:
      - multipart/form-data
      description: 注册接口
      parameters:
      - description: 账号
        in: formData
        name: email
        required: true
        type: string
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      - description: 邮件类型 1.注册 2.找回密码
        in: formData
        name: email_type
        type: integer
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 确认密码
        in: formData
        name: password_repeat
        required: true
        type: string
      - description: 验证码
        in: formData
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: Registered 注册
      tags:
      - 登录相关
  /auth/sendEmailCode:
    post:
      consumes:
      - multipart/form-data
      description: 发送邮件接口
      parameters:
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 邮件类型 1.注册 2.找回密码
        in: formData
        name: email_type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: Registered 发送邮件
      tags:
      - 登录相关
  /friends/:
    get:
      description: 获取好友列表
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/im_friends.ImFriends'
                  type: array
              type: object
      summary: friends 获取好友列表
      tags:
      - 好友
  /friends/:id:
    get:
      description: 获取好友详情
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/im_friends.ImFriends'
              type: object
      summary: friends/:id 获取好友详情
      tags:
      - 好友
  /friends/record:
    get:
      description: 获取好友申请记录
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/im_friend_records.ImFriendRecords'
                  type: array
              type: object
      summary: friends/record 获取好友申请记录
      tags:
      - 好友申请
    post:
      description: 发起好友申请
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: 添加人id
        in: formData
        name: to_id
        required: true
        type: string
      - description: 添加描述
        in: formData
        name: information
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/im_friend_records.ImFriendRecords'
                  type: array
              type: object
      summary: friends/record 发起好友申请
      tags:
      - 好友申请
    put:
      description: 同意或者拒绝请求
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: 记录id
        in: formData
        name: id
        required: true
        type: string
      - description: 状态 0等待 1同意 2拒绝
        in: formData
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: record 同意或者拒绝请求
      tags:
      - 好友申请
  /friends/status/:id:
    get:
      description: 获取好友在线状态
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/friend.UserStatus'
              type: object
      summary: friends/:id 获取好友在线状态
      tags:
      - 好友
...
swagger: "2.0"